do not delete past events on removing series from a course

+++ models/OCSeriesModel.php
@@ -149,10 +149,13 @@ class OCSeriesModel
     {
         $scheduled_episodes = self::getScheduledEpisodes($course_id);
         $scheduler_client = SchedulerClient::getInstance();
+        $time = time();
 
         foreach ($scheduled_episodes as $episode) {
-            $scheduler_client->deleteEventForSeminar($course_id, $episode['resource_id'], $episode['date_id']);
-            StudipLog::log('OC_CANCEL_SCHEDULED_EVENT', $episode['date_id'], $course_id);
+            if ($episode['start'] > $time) {
+                $scheduler_client->deleteEventForSeminar($course_id, $episode['resource_id'], $episode['date_id']);
+                StudipLog::log('OC_CANCEL_SCHEDULED_EVENT', $episode['date_id'], $course_id);
+            }
         }
 
         // Delete series from studip db
